Full refactoring: moved Convert methods; removed Rhino dependency.
