+++ Speckle_Adapter/AdapterActions/Push.cs
@@ -75,7 +75,7 @@ namespace BH.Adapter.Speckle
             try
             {
                 // Issue: with `StreamUpdateAsync` Speckle doesn't seem to send anything if the Stream is initially empty.
-                // You need to Push twice if the Stream is empty.
+                // You need to Push twice if the Stream is initially empty.
                 var updateResponse = SpeckleClient.StreamUpdateAsync(SpeckleStreamId, SpeckleStream).Result;
                 SpeckleClient.BroadcastMessage("stream", SpeckleStreamId, new { eventType = "update-global" });
             }
