+++ Speckle_Adapter/CRUD/Create/CreateIBHoMObjects.cs
@@ -45,7 +45,7 @@ namespace BH.Adapter.Speckle
             bhomObject.CustomData["Speckle_StreamId"] = SpeckleStreamId;
 
             // SpeckleObject "container". Top level has geometry representation of BHoM Object, so it can be visualised in the SpeckleViewer. 
-            SpeckleObject speckleObject = Engine.Speckle.Compute.SpeckleRepresentation(bhomObject, config.RendermeshOptions);
+            SpeckleObject speckleObject = SpeckleRepresentation(bhomObject, config.RendermeshOptions);
 
             // If no Speckle representation is found, it will be sent as an "Abstract" SpeckleObject (no visualisation).
             if (speckleObject == null)
