+++ Speckle_Adapter/Convert/ToSpeckle/Geometry/_IToSpeckle.cs
@@ -34,23 +34,24 @@ using System.Reflection;
 using BH.oM.Geometry;
 using BH.Engine.Base;
 using System.ComponentModel;
-using BH.oM.Structure.Elements;
-using BH.Engine.Structure;
 using BH.Engine.Rhinoceros;
 using BH.oM.Speckle;
 
-namespace BH.Engine.Speckle
+namespace BH.Adapter.Speckle
 {
     public static partial class Convert
     {
-        [Description("Convert BHoM Point to a Speckle Point")]
-        public static SpecklePoint ToSpeckle(this BHG.Point bhomPoint)
+        // Interface method
+        public static SpeckleObject IToSpeckle(this IGeometry iGeometry)
         {
-            if (bhomPoint == null) return default(SpecklePoint);
+            return ToSpeckle(iGeometry as dynamic);
+        }
 
-            SpecklePoint specklePoint = new SpecklePoint(bhomPoint.X, bhomPoint.Y, bhomPoint.Z);
-
-            return specklePoint;
+        // FallBack case for geometry conversions
+        private static SpeckleObject ToSpeckle(this IGeometry iGeometry)
+        {
+            // If more appropriate conversions are not found, return null
+            return null;
         }
     }
 }
